+++ src/lib.rs
@@ -1,3 +1,3 @@
-pub mod core;
+pub mod structures;
 pub mod fetching;
 pub mod processing;
