+++ src/structures/channel_collection.rs
