+++ src/structures/item_collection.rs
